+++ mls/nostr-mls-sqlite-storage/CHANGELOG.md
@@ -25,6 +25,12 @@
 
 ## Unreleased
 
+### Breaking changes
+
+- Remove group type from groups
+
+### Changed
+
 - Upgrade openmls to v0.7.0
 
 ## v0.43.0 - 2025/07/28